services/reservas_service/requirements.txt
django
djangorestframework
requests
mysqlclient
gunicorn